Add support for editing markdown files
Hopes to alleviate some pain points when it comes to editing markdown
files in Vim - that the keyboard navigation tends to get very wonky when
the text is soft-wrapped, and also the illegibility of reading prose
that stretches the full-length of a large window. Plugin Goyo restricts
the width of the text & centers it (focus mode) and pencil adjusts the
navigation key mappings to something a little more intuitive.